🎈perf(auto update): 优化自动更新的时机

+++ src/routes/blog/auto_update_build.ts
@@ -11,39 +11,43 @@ export async function get(req: IncomingMessage, res: any) {
   });
   req.on("end", async () => {
     const payload = JSON.parse(data.join(""));
-    console.log("[payload]", payload);
     res.writeHead(200, {
       "Content-Type": "application/json",
     });
+    if ((payload.ref as string).startsWith("refs/tags/v")) {
+      res.end(JSON.stringify({ code: 200 }));
+      /** 更新代码，刷新博客内容 */
+      exec(
+        "cd /root/doc/__sapper__ && " +
+          "rm -rf /root/doc/__sapper__/doc && " +
+          "git clone -b gh-pages git@github.com:2234839/doc.git && " +
+          "\\cp -rf ./doc/build/ ./ && " +
+          "pm2 reload blog",
+        { cwd: root_path },
+        (e, stdout) => {
+          if (e) {
+            console.log("[更新程序失败]", e);
+          } else {
+            console.log("[更新程序成功]");
+            exec(
+              `cd /root/doc/__sapper__/doc/ && \\cp -rf ./build/ ../build/ && pm2 reload blog`,
+              { cwd: root_path },
+              (e, stdout) => {
+                if (e) {
+                  console.log("[启动程序失败]", e);
+                } else {
+                  console.log("[启动程序成功]", stdout.toString());
+                }
+              },
+            );
+          }
+        },
+      );
+    } else {
+      console.log("[payload]", payload);
 
-    res.end(JSON.stringify({ code: 200 }));
-    /** 更新代码，刷新博客内容 */
-    exec(
-      "cd /root/doc/__sapper__ && " +
-        "rm -rf /root/doc/__sapper__/doc && " +
-        "git clone -b gh-pages git@github.com:2234839/doc.git && " +
-        "\\cp -rf ./doc/build/ ./ && " +
-        "pm2 reload blog",
-      { cwd: root_path },
-      (e, stdout) => {
-        if (e) {
-          console.log("[更新程序失败]", e);
-        } else {
-          console.log("[更新程序成功]");
-          exec(
-            `cd /root/doc/__sapper__/doc/ && \\cp -rf ./build/ ../build/ && pm2 reload blog`,
-            { cwd: root_path },
-            (e, stdout) => {
-              if (e) {
-                console.log("[启动程序失败]", e);
-              } else {
-                console.log("[启动程序成功]", stdout.toString());
-              }
-            },
-          );
-        }
-      },
-    );
+      return res.end(JSON.stringify({ code: 200 }));
+    }
   });
 }
